+++ lib/password_check.dart
@@ -1,3 +1,5 @@
 library password_check;
 
+export 'src/password_rule.dart';
+export 'src/validation_item.dart';
 export 'widgets/password_checklist.dart';